vbe: Add 'reserved' and 'oem_data' fields to vbe_info struct
We will use the additional space for storing the video mode list and the
OEM strings, making the structure self-contained.  The VBE specification
allows the implementation to use these fields in the same way; however,
since we can't rely on this behavior, we will have to copy the data to
this area.