gfx: Support actively waiting for VGA retrace
When testing the effective refresh rate suggests that the
`VBE_WAIT_FOR_RETRACE` flag has no effect, we test the refresh rate
again, but with added calls to `vga_wait_for_retrace()`.  If this works,
we set the `vsync_supported` flag and the `vsync_requires_vga_wait` flag
in `gfx_mode_info`.