scene: Add damage list system
Before this commit, we have always redrawn the complete background in
every frame, which, in our case, is not the most efficient solution.
This commit adds a damage list system, which keeps a list of "damage
rectangles" for both drawing buffers, and redraws only those parts of
the background that have been damaged in the last drawing cycle to the
respective buffer.