game: Use duck visibility test
This fixes a bug that occured when running under VirtualBox: VBox does
not emulate the timing of the vertical retrace, which results in a very
high frame rate.  As a consequence, many ducks were getting deleted
immediately after their creation because they had no time to move into
the visible area between two game updates, which the game interpreted as
having moved outside the visual area.